# docs/openapi/platform-v1.yaml
openapi: 3.1.0
info:
  title: OpenFinance Platform API
  version: 1.0.0
  description: API multiempresa para espelhos ERP, movimentos bancários, conciliação e recebimentos.
servers:
  - url: https://platform.example.com/api/v1
security:
  - bearerAuth: []
paths:
...
  /bank-connections/{id}/sync:
    post:
      operationId: syncBankConnection
      parameters:
        - { $ref: '#/components/parameters/ResourceId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
...
  /pix/charges/{id}/refunds:
    post:
      operationId: refundPix
      parameters:
        - { $ref: '#/components/parameters/ResourceId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_minor]
              properties:
                amount_minor: { $ref: '#/components/schemas/MoneyMinor' }
                external_transaction_id: { type: [string, 'null'], description: EndToEndId quando diferente do identificador da cobrança }
                refund_id: { type: [string, 'null'] }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
  /boletos/{id}:
    patch:
      operationId: updateBoleto
      parameters:
        - { $ref: '#/components/parameters/ResourceId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount_minor: { $ref: '#/components/schemas/MoneyMinor' }
                due_at: { type: string, format: date }
                payer: { type: object, additionalProperties: true }
                options: { type: object, additionalProperties: true }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
  /boletos/{id}/cancel:
    post:
      operationId: cancelBoleto
      parameters:
        - { $ref: '#/components/parameters/ResourceId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  parameters:
    ResourceId: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, maxLength: 200 } }
  responses:
    Accepted: { description: Persistido ou enfileirado com sucesso }
...
  schemas:
    MoneyMinor: { type: integer, format: int64, minimum: 0, description: Valor em unidades mínimas; centavos para BRL }
...